relocate keyboard move help controller, route and view
