--- app/controllers/KeyboardMove.scala
@@ -1,14 +1,0 @@
-package controllers
-
-import views._
-import lila.app._
-
-final class KeyboardMove(
-    env: Env
-) extends LilaController(env) {
-
-  def help =
-    Open { implicit ctx =>
-      Ok(html.keyboardMove.help()).fuccess
-    }
-}
